+++ src/config.py
@@ -1,44 +1,61 @@
 import os
+import textwrap
 
 REQUIRED_ENV_VARS = [
-    'CHANNEL_ID',
+    'SLACK_CHANNEL_ID',
     'SLACK_BOT_TOKEN',
     'SLACK_SIGNING_SECRET',
-    'ENV_VAR_3',
     'DEPLOYMENT_ENV'
 ]
 
-deployment_env = os.environ['DEPLOYMENT_ENV'].lower()
-
-match deployment_env:
-    case 'gcp':
-        REQUIRED_ENV_VARS.append('FIRESTORE_COLLECTION')
-    case 'local':
-        REQUIRED_ENV_VARS.append('STATE_FOLDER')
+GOOGLE_ENV_VARS = [
+    'GCP_REGION',
+    'GCP_SERVICE_NAME',
+    'FIRESTORE_COLLECTION'
+]
 
 
-class BaseConfig:
-    STATE_FILE: str = os.environ.get(
-        'STATE_FILE', '/mnt/persistence/exclusions.json')
-    SCHEDULE_TIMER: int = int(os.environ.get('SCHEDULE_TIMER', 60))
-    LOGGING_LEVEL = os.environ.get('LOGGING_LEVEL', "INFO").upper()
-    CHANNEL_ID = os.environ.get("CHANNEL_ID")
-    COLLECTION = os.environ.get("FIRESTORE_COLLECTION", "musical-chair-slackbot")
-    SLACK_BOT_TOKEN = os.environ.get("SLACK_BOT_TOKEN")
-    SLACK_SIGNING_SECRET = os.environ.get("SLACK_SIGNING_SECRET")
-    BUCKET = os.environ.get("GCS_BUCKET")
-    DEPLOYMENT_ENV = os.environ['DEPLOYMENT_ENV'].lower()
-
-
-def get_env_vars():
+def check_for_missing_vars(var_list):
     missing_vars = []
-    for var in REQUIRED_ENV_VARS:
+    for var in var_list:
         if var not in os.environ:
             missing_vars.append(var)
 
-
     if missing_vars:
-        raise EnvironmentError(f'''Missing required environment
-                               variable list: {missing_vars}''')
+        error_message = f'''
+            Missing required environment variables:
+            {missing_vars}
+        '''
+        raise EnvironmentError(textwrap.dedent(error_message))
 
-    return BaseConfig
+
+class BaseConfig():
+    def __init__(self):
+        check_for_missing_vars(REQUIRED_ENV_VARS)
+
+    LOGGING_LEVEL = os.environ.get('LOGGING_LEVEL', 'INFO').upper()
+    CHANNEL_ID = os.environ.get('SLACK_CHANNEL_ID')
+    COLLECTION = os.environ.get('FIRESTORE_COLLECTION', 'musical-chair-slackbot')
+    SLACK_BOT_TOKEN = os.environ.get('SLACK_BOT_TOKEN')
+    SLACK_SIGNING_SECRET = os.environ.get('SLACK_SIGNING_SECRET')
+    BUCKET = os.environ.get('GCS_BUCKET')
+    PORT = os.environ.get('PORT', 8000)
+    DEPLOYMENT_ENV = 'local'
+
+
+class GcpConfig(BaseConfig):
+    def __init__(self):
+        super().__init__()
+        check_for_missing_vars(GOOGLE_ENV_VARS)
+
+    REGION = os.getenv("GCP_REGION", "europe-west2")
+    SERVICE_NAME = os.getenv("GCP_SERVICE_NAME", "musical-chair-cr")
+    DEPLOYMENT_ENV = 'GCP'
+
+
+def get_env_vars():
+    deployment_env = os.environ.get('DEPLOYMENT_ENV', 'GCP').lower()
+    if deployment_env == 'local':
+        return BaseConfig
+
+    return GcpConfig
