+++ src/features/__init__.py
@@ -1,0 +1,8 @@
+from .ask_for_song.endpoint import router as ask_for_song_router
+from .playlist_handler.endpoint import router as playlist_router
+from .playlist_handler.endpoint import handle_song_url
+
+__all__ = [
+    "ask_for_song_router",
+    "playlist_router",
+]
