+++ src/slack_events/app_home_opened.py
@@ -1,0 +1,42 @@
+from app_instances import slack_app
+
+@slack_app.event('app_home_opened')
+async def update_home_tab(client, event, logger):
+    user_id = event['user']
+    await client.views_publish(
+        user_id=user_id,
+        view={
+            "type": "home",
+            "blocks": [
+                {
+                    "type": "header",
+                    "text": {
+                        "type": "plain_text",
+                        "text": "Musical Chair",
+                    }
+                },
+                {
+                    "type": "section",
+                    "text": {
+                        "type": "mrkdwn",
+                        "text": f"Hello <@{user_id}>\n\nWelcome to the home page for the sometimes not-so-friendly music bot :musical_note::chair:"
+                    }
+                },
+                {
+                    "type": "actions",
+                    "elements": [
+                        {
+                            "type": "button",
+                            "text": {
+                                "type": "plain_text",
+                                "text": "Opt-Out",
+                                "emoji": True
+                            },
+                            "value": "click_me_123",
+                            "url": "https://google.com"
+                        }
+                    ]
+                }
+            ]
+        }
+    )
